test: migrate stats.rs unit tests to integration tests
- Migrated all 36 tests from src/authorship/stats.rs to tests/integration/stats_unit.rs
- Made 3 helper functions public (accepted_lines_from_attestations, line_range_overlap_len, calculate_waiting_time)
- Converted TmpRepo tests to use TestRepo with CLI-based git operations
- Moved and renamed 14 insta snapshot files to match new module path
- All tests pass including 1 ignored test (test_stats_from_authorship_log_mixed_cap)
