+++ tests/integration/main.rs
@@ -98,6 +98,7 @@ mod stale_prompt_carry;
 mod stash_attribution;
 mod stash_hooks_unit;
 mod stats;
+mod stats_unit;
 mod status_ignore;
 mod status_unit;
 mod subdirs;
